Update README.md
This project is a simple tool for learning and working on programming projects. Using various languages like Python, JavaScript, and C, users can strengthen their programming skills and easily write, modify, and create small projects. The goal of this project is to simplify the programming process and create an environment for better learning.